backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
Django==3.2.7
psycopg2-binary>=2.8
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
fonttools==4.27.0
html5lib==1.1
idna==3.2
kombu==5.1.0
Pillow==8.3.2
prompt-toolkit==3.0.20
pycparser==2.20
pydyf==0.1.1
PyJWT==2.1.0
pyphen==0.11.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.0
twilio==7.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
weasyprint==53.3
webencodings==0.5.1
zopfli==0.1.8
